Fix draggable view relative position bugs
Draggable view was not expanding and collapsing in the same position relative to screen size as expected. Update positioning logic to use relative screensize instead of just the view center to achieve desired results.

Still needs to be updated to support more modularity and customization and less hardcoded padding values.

Move logic for collapsing into the draggable view and out of the view controller where it didn't belong.

Update draggable view to have padding on the bottom of the view to support the spring animation and prevent showing background when "bouncing".

General code cleanup and commenting.

Resolves: #TicketNumber